// src/types/classroom-type.ts
export  type classroomType = {
    class_id: number;
    class_name: string;
    user_id: number;
    description: string;
    code: string;
    color_id: number;
    color_name: string;
    subject_id: number;
    subject_name: string;
    subject_type_id: number;
    subject_type_name: string;
    status_id:number;
    status_name:string;
    permission_id: number;
    permission_name: string;
    teacher_name: string;
    user_name: string;
    created_at: string;
    updated_at: string;
    student_count: number

};